requirements.txt
apache-airflow==2.7.0
requests==2.31.0
pandas==2.1.1
SQLAlchemy==2.1.0
python-dotenv==1.0.0
